fix serious bug on imposition
because we were searching for folds the size of the signature,
unexpected behaviour ocurred when the total pages were smaller than the
signature, leading to missing pages on the final result.

halving by the fold size allows to halve by the signature on full folds
and by the amount of trailing pages on the incomplete fold.